Simplify bb8 edit function
switched to reading the file as JSON and updating
the genes dict instead of search and replacing
values in the bytes.

The previous approach was used because save files
were being corrupted, but it turns out this had
to do with how the zipfile was being created, not
how the individual bb8 files were encoded.